Create README.md
#  Work Day Scheduler

## Purpose 
"Work Day Scheduler" is a simple calendar app for scheduling your productive day.

To make people's lives easy and more manageable. Sometimes it can be overwhelming and tasking to list everything you need to do that day in your head. This application helps you put everything out, it can be edit at any time and saved. Once the task is in grey it is completed, red means you are working on it, and green means you will take care of it in the later hours to come. 



## Built With

  * HTML
  * CSS
  * JavaScript
  * jQuery
  * Moment.js
  * Bootstrap


## Authors
fatou fall




Copyright (c) 2021 Fatou fall